+++ .changeset/rare-oranges-flash.md
@@ -1,0 +1,5 @@
+---
+"claude-dev": patch
+---
+
+Add "Use custom prompt" option to Ollama provider
